Revert "38 design 검색 리스트 UI 작업 (#45)"
This reverts commit 43737f212de7a3b23987c84327cf55b99abaac64.
