--- pages/product/index.tsx
@@ -1,7 +1,0 @@
-import { ProductPage } from '@pages/product';
-
-const Product = () => {
-  return <ProductPage />;
-};
-
-export default Product;
