+++ src/entities/product/product-card-item/ui/description.tsx
@@ -1,4 +1,4 @@
-import { formatPrice } from '@shared/utils/price';
+import { formatPrice } from '@shared/utils/format-price';
 
 import * as S from './product-card-item-style';
