--- src/pages/product/index.ts
@@ -1,1 +1,0 @@
-export { ProductPage } from './product-page';
