--- src/pages/product/ui/card-list-with-sorting-box.tsx
@@ -1,7 +1,0 @@
-import { ProductList } from '@widgets/recent-product-list';
-
-const CardListWithSortingBox = () => {
-  return <ProductList />;
-};
-
-export { CardListWithSortingBox };
