--- src/shared/ui/category-filter/index.ts
@@ -1,1 +1,0 @@
-export { CategoryFilter } from './category-filter';
