--- src/shared/ui/checkbox/index.ts
@@ -1,1 +1,0 @@
-export { CommonCheckBox } from './common-checkbox/common-checkbox';
