+++ src/shared/ui/inputs/serch-input/search-input.stories.ts
@@ -15,8 +15,5 @@ export const Default: Story = {
   args: {
     placeholder: '원하시는 상품을 검색해보세요!',
     maxWidth: '100%',
-    handleSubmit(e) {
-      console.log(e);
-    },
   },
 };
