+++ src/shared/ui/inputs/serch-input/search-input.tsx
@@ -1,5 +1,3 @@
-import { FormEvent, LegacyRef } from 'react';
-
 import Image from 'next/image';
 
 import * as S from './search-input-style';
@@ -7,18 +5,14 @@ import * as S from './search-input-style';
 interface SearchInputProps {
   placeholder: string;
   maxWidth?: string;
-  handleSubmit: (e: FormEvent<HTMLFormElement>) => void;
-  inputRef?: LegacyRef<HTMLInputElement>;
 }
 
-const SearchInput = ({ handleSubmit, inputRef, placeholder, maxWidth = '100%' }: SearchInputProps) => {
+const SearchInput = ({ placeholder, maxWidth = '100%' }: SearchInputProps) => {
   return (
     <>
       <S.SearchInputBox $maxWidth={maxWidth}>
-        <form onSubmit={handleSubmit}>
-          <Image width={19} height={19} src='/images/icon/search-icon.svg' alt='검색 아이콘' />
-          <input ref={inputRef} type='text' placeholder={placeholder} />
-        </form>
+        <Image width={19} height={19} src='/images/icon/search-icon.svg' alt='검색 아이콘' />
+        <input type='text' placeholder={placeholder} />
       </S.SearchInputBox>
     </>
   );
