--- src/shared/ui/select-box/index.ts
@@ -1,1 +1,0 @@
-export { CommonSelectBox } from './common-select-box/common-select-box';
