--- src/shared/ui/select-box/lib/product-list-sorting-data.ts
@@ -1,24 +1,0 @@
-import { DropDownBoxDto } from '../ui/drop-down-box';
-
-export const productListSortingData: DropDownBoxDto[] = [
-  {
-    id: 0,
-    name: '최신순',
-  },
-  {
-    id: 1,
-    name: '조회순',
-  },
-  {
-    id: 2,
-    name: '인기순',
-  },
-  {
-    id: 3,
-    name: '낮은 가격순',
-  },
-  {
-    id: 4,
-    name: '높은 가격순',
-  },
-];
