+++ src/shared/ui/styles/utills/index.ts
@@ -3,5 +3,3 @@ export { inputReset, inputContainer } from './input-reset';
 export { SwiperPaginationReset } from './swiper-reset';
 
 export { textLineClamp2, backgroundClip } from './text-style';
-
-export { dropDownMenuStyle } from './drop-down-menu';
