+++ src/shared/utils/format-price/format-price.ts
@@ -1,0 +1,5 @@
+const formatPrice = (price: number) => {
+  return price.toLocaleString();
+};
+
+export { formatPrice };
