+++ src/shared/utils/format-price/index.ts
