+++ src/shared/utils/price/index.ts
@@ -1,3 +1,1 @@
 export { formatPriceToKoStyle } from './format-price-to-ko-style';
-
-export { formatPrice } from './format-price';
