--- src/widgets/product-category-filter/index.ts
@@ -1,1 +1,0 @@
-export { ProductCategoryFilterBox } from './ui/product-category-filter';
