--- src/widgets/recent-product-list/ui/product-list-style.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-interface ProductListProps {
-  $cardListFlexGap: string;
-}
-
-export const ProductListContainer = styled.div<ProductListProps>`
-  display: flex;
-  flex-wrap: wrap;
-  gap: ${(props) => props.$cardListFlexGap};
-  align-items: center;
-`;
